Split tests of datatransfer into unit and integration
