+++ softlab-data-transfer/src/integrationTest/kotlin/org/softlab/datatransfer/adapters/AdapterProviderIT.kt
@@ -1,0 +1,65 @@
+package org.softlab.datatransfer.adapters
+
+import org.junit.jupiter.api.BeforeAll
+import org.junit.jupiter.api.Test
+import org.softlab.datataset.test.initiators.createMongoContainer
+import org.softlab.datataset.test.initiators.createPostgresContainer
+import org.softlab.datatransfer.adapters.mongo.MongoDestination
+import org.softlab.datatransfer.adapters.mongo.MongoSource
+import org.softlab.datatransfer.adapters.postgres.PostgresDestination
+import org.softlab.datatransfer.adapters.postgres.PostgresSource
+import org.testcontainers.junit.jupiter.Container
+import org.testcontainers.junit.jupiter.Testcontainers
+import kotlin.test.assertIs
+
+
+@Testcontainers
+class AdapterProviderIT {
+    companion object {
+        @Container
+        @JvmStatic
+        private val postgres = createPostgresContainer()
+
+        @Container
+        @JvmStatic
+        private val mongo = createMongoContainer()
+
+        private lateinit var postgresUri: String
+        private lateinit var mongoUri: String
+
+        @BeforeAll
+        @JvmStatic
+        fun setup() {
+            postgresUri = "${postgres.jdbcUrl}&user=${postgres.username}&password=${postgres.password}"
+            mongoUri = "${mongo.connectionString}/testdb"
+        }
+    }
+
+    @Test
+    fun `sourceFor() returns postgres source for postgres URI`() {
+        AdapterProvider.sourceFor(postgresUri).use { source ->
+            assertIs<PostgresSource>(source)
+        }
+    }
+
+    @Test
+    fun `sourceFor() returns mongo source for mongo URI`() {
+        AdapterProvider.sourceFor(mongoUri).use { source ->
+            assertIs<MongoSource>(source)
+        }
+    }
+
+    @Test
+    fun `destinationFor() returns postgres destination for postgres URI`() {
+        AdapterProvider.destinationFor(postgresUri).use { destination ->
+            assertIs<PostgresDestination>(destination)
+        }
+    }
+
+    @Test
+    fun `destinationFor() returns mongo destination for mongo URI`() {
+        AdapterProvider.destinationFor(mongoUri).use { destination ->
+            assertIs<MongoDestination>(destination)
+        }
+    }
+}
