+++ softlab-data-transfer/src/test/kotlin/org/softlab/datatransfer/adapters/AdapterProviderTest.kt
@@ -2,58 +2,13 @@ package org.softlab.datatransfer.adapters
 
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.containsString
-import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import org.softlab.datataset.test.initiators.createMongoContainer
-import org.softlab.datataset.test.initiators.createPostgresContainer
-import org.softlab.datatransfer.adapters.mongo.MongoDestination
-import org.softlab.datatransfer.adapters.mongo.MongoSource
-import org.softlab.datatransfer.adapters.postgres.PostgresDestination
-import org.softlab.datatransfer.adapters.postgres.PostgresSource
-import org.testcontainers.junit.jupiter.Container
-import org.testcontainers.junit.jupiter.Testcontainers
-import kotlin.test.assertIs
 
 
-@Testcontainers
 class AdapterProviderTest {
-    companion object {
-        @Container
-        @JvmStatic
-        private val postgres = createPostgresContainer()
-
-        @Container
-        @JvmStatic
-        private val mongo = createMongoContainer()
-
-        private lateinit var postgresUri: String
-        private lateinit var mongoUri: String
-
-        @BeforeAll
-        @JvmStatic
-        fun setup() {
-            postgresUri = "${postgres.jdbcUrl}&user=${postgres.username}&password=${postgres.password}"
-            mongoUri = "${mongo.connectionString}/testdb"
-        }
-    }
-
     @Test
-    fun `sourceFor() returns postgres source for postgres URI`() {
-        AdapterProvider.sourceFor(postgresUri).use { source ->
-            assertIs<PostgresSource>(source)
-        }
-    }
-
-    @Test
-    fun `sourceFor() returns mongo source for mongo URI`() {
-        AdapterProvider.sourceFor(mongoUri).use { source ->
-            assertIs<MongoSource>(source)
-        }
-    }
-
-    @Test
-    fun `sourceFor() throws for unsupported URI`() {
+    fun `sourceFor() throws for unknown URI`() {
         val exc = assertThrows<IllegalStateException> {
             AdapterProvider.sourceFor("unknown://db")
         }
@@ -61,21 +16,7 @@ class AdapterProviderTest {
     }
 
     @Test
-    fun `destinationFor() returns postgres destination for postgres URI`() {
-        AdapterProvider.destinationFor(postgresUri).use { destination ->
-            assertIs<PostgresDestination>(destination)
-        }
-    }
-
-    @Test
-    fun `destinationFor() returns mongo destination for mongo URI`() {
-        AdapterProvider.destinationFor(mongoUri).use { destination ->
-            assertIs<MongoDestination>(destination)
-        }
-    }
-
-    @Test
-    fun `destinationFor() throws for unsupported URI`() {
+    fun `destinationFor() throws for unknown URI`() {
         val exc = assertThrows<IllegalStateException> {
             AdapterProvider.destinationFor("unknown://db")
         }
